Melhoria na escrita do codigo e layout
