+++ app/src/main/java/com/jess/eaiclubnasa/Constants.kt
@@ -2,5 +2,6 @@ package com.jess.eaiclubnasa
 
 object Constants {
     const val API_KEY = "2OYIJlQnpyYdlDogabi1iJnq5b4HRMXdtPwA3e4O"
-    const val DATE_PATTERN = "yyyy-MM-dd"
+    const val DATE_PATTERN_YEAR_MONTH_DAY = "yyyy-MM-dd"
+    const val DATE_PATTERN_DAY_MONTH_YEAR = "dd/MM/yyyy"
 }
