+++ app/src/main/java/com/jess/eaiclubnasa/view/ApodAdapter.kt
@@ -7,6 +7,7 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
+import com.jess.eaiclubnasa.ApodUtils.formatDate
 import com.jess.eaiclubnasa.R
 import com.jess.eaiclubnasa.model.ApodResult
 
@@ -30,11 +31,13 @@ class ApodAdapter(var listApod: MutableList<ApodResult>) :
     class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
         private var apodImage = itemView.findViewById<ImageView>(R.id.iv_apod)
         private var apodTitle = itemView.findViewById<TextView>(R.id.tv_apod)
+        private var apodDate = itemView.findViewById<TextView>(R.id.tv_apod_date)
 
         fun onBind(apodResult: ApodResult) {
             Glide.with(itemView.context).load(apodResult.url).placeholder(R.drawable.logo)
                 .into(apodImage)
             apodTitle.text = apodResult.title
+            apodDate.text = formatDate(apodResult.date)
         }
     }
 
@@ -49,5 +52,4 @@ class ApodAdapter(var listApod: MutableList<ApodResult>) :
     }
 
     fun getLastItem() = listApod.last()
-
 }
