+++ app/src/main/res/layout/item_apod_recyclerview.xml
@@ -30,13 +30,26 @@
             android:layout_marginTop="@dimen/margin_16"
             android:layout_marginEnd="@dimen/margin_8"
             android:layout_marginRight="@dimen/margin_8"
-            android:layout_marginBottom="@dimen/margin_16"
+            android:textColor="@color/yellow_ffb74d"
             android:textSize="@dimen/text_size_14"
-            app:layout_constraintBottom_toBottomOf="parent"
+            android:textStyle="bold"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/iv_apod"
             tools:text="Astronomy Picture of the Day" />
+
+        <TextView
+            android:id="@+id/tv_apod_date"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="@dimen/margin_8"
+            android:layout_marginBottom="@dimen/margin_16"
+            android:textColor="@color/yellow_ffb74d"
+            android:textStyle="bold"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toStartOf="@+id/tv_apod"
+            app:layout_constraintTop_toBottomOf="@+id/tv_apod"
+            tools:text="10/11/2020" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </androidx.cardview.widget.CardView>
